app/requirements.txt
Django==3.1.1
djangorestframework==3.11.2
djangorestframework-jsonapi==3.2.0
django-filter==2.4.0
